refactor: 💡 About route goes to Home
