+++ src/components/molecules/About.pcss
@@ -1,5 +1,5 @@
 .about-wrapper {
-  @apply max-w-[80%] mx-auto text-xl;
+  @apply mt-24 text-base text-justify;
 
   p {
     @apply mb-8;
@@ -7,5 +7,9 @@
 
   a {
     @apply font-bold underline hover:no-underline;
+  }
+
+  .capitular {
+    @apply text-2xl;
   }
 }
