+++ src/components/molecules/About.tsx
@@ -1,3 +1,4 @@
+import Title from "../atoms/Title";
 import "./About.pcss";
 
 // TODO: Write text with Markdown, loading package on Suspense
@@ -5,12 +6,15 @@ import "./About.pcss";
 const About = () => {
   return (
     <div className="about-wrapper">
+      <Title>Sobre</Title>
+
       <p>
-        <strong className="text-3xl">Sextou!</strong> é uma plataforma criada
+        <strong className="capitular">Sextou!</strong> é uma plataforma criada
         para empoderar o usuário na escolha de seu entretenimento, encontrando
         novas opções e filtrando a que melhor se encaixa em seu gosto e
         localização.
       </p>
+
       <p>
         O motivo de criação da plataforma é facilitar a divulgação e encontro de
         eventos que acontecem, principalmente fora do centro cultural do estado
@@ -18,11 +22,13 @@ const About = () => {
         Eventos do Facebook, após a migração de usuários para fora da
         plataforma.
       </p>
+
       <p>
         Plataforma idealizada e criada por mim,{" "}
         <a href="https://linkedin.com/in/didiraja" target="_blank">
           Dico Didiraja
         </a>
+        .
       </p>
     </div>
   );
