+++ src/routes/index.tsx
@@ -1,7 +1,6 @@
 import { createBrowserRouter } from "react-router-dom";
 
 import LayoutHome from "../components/organisms/Layout.Home";
-import About from "./About";
 import Category from "./Category";
 import NotFound from "./NotFound";
 
@@ -10,10 +9,6 @@ const router = createBrowserRouter([
     path: "/",
     element: <LayoutHome />,
     errorElement: <NotFound />,
-  },
-  {
-    path: "/about",
-    element: <About />,
   },
   {
     path: "/category/:slug",
